Post and display tutor reviews

+++ backend/app/models.py
@@ -95,7 +95,7 @@ class Review(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     student_id = db.Column(db.Integer, db.ForeignKey('student.id'), nullable=False)
     tutor_id = db.Column(db.Integer, db.ForeignKey('tutor.id'), nullable=False)
-    rating = db.Column(db.Integer, nullable=False)
+    rating = db.Column(db.Float, nullable=False)
     description = db.Column(db.String(255), nullable=False)
 
     def __repr__(self):
